requirements.txt
streamlit==1.37.1
faiss-cpu==1.8.0.post1
pypdf==5.0.1
scikit-learn==1.5.1
numpy<2.0